refactor: Refactor chunked_data_view tests
