+++ tests/chunked_data_view/CMakeLists.txt
@@ -2,7 +2,6 @@ list(APPEND tests
     view
     axis
     index_mapper
-    builder
     request_manipulation_bounding_box
     bounding_box
     view_individual_chunking
